Major Refactoring - Moved API connection into ConnectionHandler.java - Moved Ping command into new package
