+++ src/main/java/com/Pieman492/PieCannon/core/Piecannon.java
@@ -1,0 +1,14 @@
+package com.Pieman492.PieCannon.core;
+
+import com.Pieman492.PieCannon.commands.*;
+
+public class Piecannon {
+
+    public static void main(String[] args) {
+
+        ConnectionHandler connectionHandler = new ConnectionHandler();
+        new Annoy(connectionHandler.getClient());
+        new Ping(connectionHandler.getClient());
+
+    }
+}
